Gen7\Wild7Randomizer -Added new randomization methods for SOS slots, Day/Night tables.  Gives a more even distribution rate for all species -Added option to balance encounter rates.  Allows for easier hunting
Gen7\PersonalEditor7.cs
-Added GuaranteedItemChance flag.  Guarantees all wild species carry items when checked.
-Added Catch Rate, Call Rate, Friendship, Gender,  and Color to Log file Dump
-Resized, added, and moved controls on form respectively.

Gen7\SMWE.cs
-Redesigned Summary for Full and Simple options.  Full will show full stats for each encounter slot and each SOS under each one, including stats and held items.
-ExecuteRandomization: Added missing Shedinja flag and AlwaysShuffle flag to speciesRandomizer constructor

PK3DS.Core\Randomizers\GenericRandomizer.cs
-Added Debug.print

PK3DS.Core\Randomizers\PersonalRandomizer.cs
-Added GuaranteedItemChance flag logic
-RandomizeHeldItems: implemented randomization for all 3 hold chance slots

PK3DS.Core\Randomizers\SpeciesRandomizer.cs
-Added AlwaysShuffle flag and logic.  Flag gives option to allow repeats of randomized species

PK3DS.Core\Structures\Gen7\Area7.cs
-Added GetSummaryX.  I think all this does is change the sort order of tables from GetSummary.

PK3DS.Core\Structures\Gen7\Encounter7.cs
-Added spacing to string for formatting/readability purposes

PK3DS.Core\Structures\Gen7\EncounterTable.cs
-Maded GetSlotSetSummary public.  It's useful in  log dumps.

+++ pk3DS/Subforms/Gen6/MegaEvoEditor6.cs
@@ -132,7 +132,7 @@ namespace pk3DS
             {
                 if (me.Method[i] > 1)
                     return; // Shouldn't hit this.
-                me.Method[i] = checkbox_spec[i].Checked ? (byte)1 : (byte)0;
+                me.Method[i] = checkbox_spec[i].Checked ? 1 : 0;
                 me.Argument[i] = (ushort)WinFormsUtil.GetIndex(item_spec[i]);
                 me.Form[i] = (ushort)forme_spec[i].SelectedIndex;
             }
